+++ package/src/functions/rowid.rs
@@ -1,8 +1,11 @@
 use std::time::SystemTime;
 
-use crate::functions::{
-    encode::{EncodeOptions, encode_unsafe},
-    get_randomness::{_get_randomness, GetRandomnessOptions},
+use crate::{
+    functions::{
+        encode::{EncodeOptions, encode_unsafe},
+        get_randomness::{_get_randomness, GetRandomnessOptions},
+    },
+    time::system_time_to_timestamp,
 };
 
 pub struct RowIDOptions<CharList: AsRef<str>> {
@@ -13,9 +16,11 @@ pub struct RowIDOptions<CharList: AsRef<str>> {
 pub fn _rowid<CharList: AsRef<str>>(opts: RowIDOptions<CharList>) -> String {
     let char_list: &str = opts.char_list.as_ref();
 
-    encode_unsafe(EncodeOptions { char_list, system_time: SystemTime::now() })
-        + &_get_randomness(GetRandomnessOptions {
-            char_list,
-            randomness_length: opts.randomness_length,
-        })
+    encode_unsafe(EncodeOptions {
+        char_list,
+        timestamp: system_time_to_timestamp(SystemTime::now()),
+    }) + &_get_randomness(GetRandomnessOptions {
+        char_list,
+        randomness_length: opts.randomness_length,
+    })
 }