+++ package/src/with_config.rs
@@ -1,4 +1,4 @@
-use std::{io, time::SystemTime};
+use std::io;
 
 use crate::{
     common::{
@@ -52,6 +52,7 @@ impl RowIDWithConfigResult {
     ///
     /// let rwc: RowIDWithConfigResult =
     ///     RowIDWithConfig::new().done().unwrap();
+    ///
     /// let id: String = rwc.rowid();
     /// ```
     pub fn rowid(&self) -> String {
@@ -68,20 +69,23 @@ impl RowIDWithConfigResult {
     ///
     /// ```no_run
     /// use std::time::SystemTime;
-    /// use rowid::with_config::{RowIDWithConfig, RowIDWithConfigResult};
+    /// use rowid::{
+    ///     with_config::{RowIDWithConfig, RowIDWithConfigResult},
+    ///     time::system_time_to_timestamp,
+    /// };
     ///
     /// let rwc: RowIDWithConfigResult =
     ///     RowIDWithConfig::new().done().unwrap();
-    /// let encoded: String = rwc.encode(SystemTime::now()).unwrap();
+    ///
+    /// let now: usize = system_time_to_timestamp(SystemTime::now());
+    ///
+    /// let encoded: String = rwc.encode(now).unwrap();
     /// ```
-    pub fn encode<T: Into<SystemTime>>(
+    pub fn encode(
         &self,
-        system_time: T,
+        timestamp: usize,
     ) -> io::Result<String> {
-        _encode(EncodeOptions {
-            char_list: &self.state.char_list,
-            system_time: system_time.into(),
-        })
+        _encode(EncodeOptions { char_list: &self.state.char_list, timestamp })
     }
 
     /// This function decodes the ID into a timestamp in milliseconds.
@@ -94,12 +98,13 @@ impl RowIDWithConfigResult {
     ///
     /// let rwc: RowIDWithConfigResult =
     ///     RowIDWithConfig::new().done().unwrap();
-    /// let decoded: SystemTime = rwc.decode("ABC123").unwrap();
+    ///
+    /// let decoded: usize = rwc.decode("ABC123").unwrap();
     /// ```
     pub fn decode<S: AsRef<str>>(
         &self,
         encoded: S,
-    ) -> io::Result<SystemTime> {
+    ) -> io::Result<usize> {
         _decode(DecodeOptions {
             char_list: &self.state.char_list,
             encoded: encoded.as_ref(),
@@ -114,22 +119,25 @@ impl RowIDWithConfigResult {
     /// use std::time::SystemTime;
     /// use rowid::{
     ///     base::GenerateResult,
-    ///     with_config::{RowIDWithConfig, RowIDWithConfigResult}
+    ///     with_config::{RowIDWithConfig, RowIDWithConfigResult},
+    ///     time::system_time_to_timestamp,
     /// };
     ///
     /// let rwc: RowIDWithConfigResult =
     ///     RowIDWithConfig::new().done().unwrap();
-    /// let now: SystemTime = SystemTime::now();
+    ///
+    /// let now: usize = system_time_to_timestamp(SystemTime::now());
+    ///
     /// let result: GenerateResult = rwc.generate(now, Some(22));
     /// ```
-    pub fn generate<T: Into<SystemTime>>(
+    pub fn generate(
         &self,
-        system_time: T,
+        timestamp: usize,
         randomness_length: Option<usize>,
     ) -> GenerateResult {
         _generate(GenerateOptions {
             char_list: &self.state.char_list,
-            system_time: system_time.into(),
+            timestamp,
             randomness_length: match randomness_length {
                 | Some(l) => l,
                 | None => self.state.randomness_length,
@@ -149,6 +157,7 @@ impl RowIDWithConfigResult {
     ///
     /// let rwc: RowIDWithConfigResult =
     ///     RowIDWithConfig::new().done().unwrap();
+    ///
     /// let result: VerifyResult = rwc.verify("ABC123");
     /// ```
     pub fn verify<S: AsRef<str>>(
@@ -170,6 +179,7 @@ impl RowIDWithConfigResult {
     ///
     /// let rwc: RowIDWithConfigResult =
     ///     RowIDWithConfig::new().done().unwrap();
+    ///
     /// let randomness: String = rwc.get_randomness(10);
     /// ```
     pub fn get_randomness(
@@ -282,6 +292,7 @@ impl RowIDWithConfig {
     ///
     /// let rwc: RowIDWithConfigResult =
     ///     RowIDWithConfig::new().done().unwrap();
+    ///
     /// let id: String = rwc.rowid();
     /// ```
     pub fn done(self) -> io::Result<RowIDWithConfigResult> {