+++ tests/src/base.rs
@@ -5,7 +5,7 @@ use rowid::{
         GenerateResult, RowIDError, VerifyResult, decode, encode, generate,
         get_randomness, rowid, verify,
     },
-    time::{system_time_to_timestamp, timestamp_to_system_time},
+    time::system_time_to_timestamp,
 };
 
 // get_randomness
@@ -38,13 +38,14 @@ fn test_get_randomness_20() {
 
 #[test]
 fn test_encode_now() {
-    let encoded: String = encode(SystemTime::now()).unwrap();
+    let encoded: String =
+        encode(system_time_to_timestamp(SystemTime::now())).unwrap();
     assert!(encoded.len() == 10);
 }
 
 #[test]
 fn test_encode_0() {
-    let encoded: String = encode(timestamp_to_system_time(0)).unwrap();
+    let encoded: String = encode(0).unwrap();
     assert!(encoded == "0000000000");
 }
 
@@ -60,11 +61,9 @@ fn test_rowid() {
 
 #[test]
 fn test_decode() {
-    let current: SystemTime = SystemTime::now();
-    let decoded: SystemTime = decode(&encode(current).unwrap()).unwrap();
-    assert!(
-        system_time_to_timestamp(decoded) == system_time_to_timestamp(current)
-    );
+    let now: usize = system_time_to_timestamp(SystemTime::now());
+    let decoded: usize = decode(&encode(now).unwrap()).unwrap();
+    assert!(decoded == now);
 }
 
 #[test]
@@ -93,14 +92,11 @@ fn test_decode_invalid_input_error() {
 
 #[test]
 fn test_generate() {
-    let current: SystemTime = SystemTime::now();
-    let generated: GenerateResult = generate(current, Some(6));
+    let now: usize = system_time_to_timestamp(SystemTime::now());
+    let generated: GenerateResult = generate(now, Some(6));
     let id: String = generated.result.unwrap();
     assert!(generated.success == true);
-    assert!(
-        system_time_to_timestamp(decode(&id).unwrap())
-            == system_time_to_timestamp(current)
-    );
+    assert!(decode(&id).unwrap() == now);
     assert!(id.len() == 16);
 }
 
@@ -108,13 +104,12 @@ fn test_generate() {
 
 #[test]
 fn test_verify() {
-    let current: SystemTime = SystemTime::now();
-    let id: String = generate(current, Some(6)).result.unwrap();
+    let now: usize = system_time_to_timestamp(SystemTime::now());
+    let id: String = generate(now, Some(6)).result.unwrap();
     let verified: VerifyResult = verify(&id);
     assert!(verified.success == true);
     assert!(match verified.result {
-        | Some(r) =>
-            system_time_to_timestamp(r) == system_time_to_timestamp(current),
+        | Some(r) => r == now,
         | None => false,
     });
     assert!(verified.natural == Some(true));