+++ tests/src/with_config.rs
@@ -22,7 +22,8 @@ fn test_rowid_with_config_encode() {
     let rwc: RowIDWithConfigResult =
         RowIDWithConfig::new().randomness_length(6).done().unwrap();
 
-    let encoded: String = rwc.encode(SystemTime::now()).unwrap();
+    let encoded: String =
+        rwc.encode(system_time_to_timestamp(SystemTime::now())).unwrap();
 
     assert!(encoded.len() == 10);
 }
@@ -42,13 +43,10 @@ fn test_rowid_with_config_decode() {
     let rwc: RowIDWithConfigResult =
         RowIDWithConfig::new().randomness_length(6).done().unwrap();
 
-    let current: SystemTime = SystemTime::now();
-    let decoded: SystemTime =
-        rwc.decode(&rwc.encode(current).unwrap()).unwrap();
+    let now: usize = system_time_to_timestamp(SystemTime::now());
+    let decoded: usize = rwc.decode(&rwc.encode(now).unwrap()).unwrap();
 
-    assert!(
-        system_time_to_timestamp(decoded) == system_time_to_timestamp(current)
-    );
+    assert!(decoded == now);
 }
 
 #[test]
@@ -59,13 +57,13 @@ fn test_rowid_with_config_encode_decode() {
         .done()
         .unwrap();
 
-    let now: SystemTime = SystemTime::now();
+    let now: usize = system_time_to_timestamp(SystemTime::now());
 
     let encoded: String = rwc.encode(now).unwrap();
 
-    let decoded: SystemTime = rwc.decode(encoded).unwrap();
+    let decoded: usize = rwc.decode(encoded).unwrap();
 
-    assert!(system_time_to_timestamp(decoded) == system_time_to_timestamp(now));
+    assert!(decoded == now);
 }
 
 #[test]
@@ -73,16 +71,13 @@ fn test_rowid_with_config_generate() {
     let rwc: RowIDWithConfigResult =
         RowIDWithConfig::new().randomness_length(6).done().unwrap();
 
-    let current: SystemTime = SystemTime::now();
-    let generated: GenerateResult = rwc.generate(current, None);
+    let now: usize = system_time_to_timestamp(SystemTime::now());
+    let generated: GenerateResult = rwc.generate(now, None);
     let id: String = generated.result.unwrap();
 
     assert!(generated.success == true);
     assert!(id.len() == 16);
-    assert!(
-        system_time_to_timestamp(rwc.decode(&id).unwrap())
-            == system_time_to_timestamp(current)
-    );
+    assert!(rwc.decode(&id).unwrap() == now);
 }
 
 #[test]
@@ -90,14 +85,13 @@ fn test_rowid_with_config_verify() {
     let rwc: RowIDWithConfigResult =
         RowIDWithConfig::new().randomness_length(6).done().unwrap();
 
-    let current: SystemTime = SystemTime::now();
-    let generated: GenerateResult = rwc.generate(current, None);
+    let now: usize = system_time_to_timestamp(SystemTime::now());
+    let generated: GenerateResult = rwc.generate(now, None);
     let verified: VerifyResult = rwc.verify(&generated.result.unwrap());
 
     assert!(verified.success == true);
     assert!(match verified.result {
-        | Some(r) =>
-            system_time_to_timestamp(r) == system_time_to_timestamp(current),
+        | Some(r) => r == now,
         | None => false,
     });
     assert!(verified.natural == Some(true));